feat: add text adapter
