+++ Casbin/Persist/Adapter/File/FileFilteredAdapter.cs
@@ -1,11 +1,16 @@
-﻿namespace Casbin.Persist.Adapter.File;
+﻿using System;
 
+namespace Casbin.Persist.Adapter.File;
+
+[Obsolete("Please use FileAdapter instead")]
 public class FileFilteredAdapter : FileAdapter
 {
+    [Obsolete("Please use FileAdapter instead")]
     public FileFilteredAdapter(string filePath) : base(filePath)
     {
     }
 
+    [Obsolete("Please use StreamAdapter instead")]
     public FileFilteredAdapter(System.IO.Stream originalStream) : base(originalStream)
     {
     }
