+++ Casbin/Persist/Adapter/Text/TextAdapter.cs
@@ -1,0 +1,9 @@
+﻿namespace Casbin.Persist.Adapter.Text;
+
+public class TextAdapter : BaseAdapter, IEpochAdapter, IFilteredAdapter
+{
+    public TextAdapter(string text)
+    {
+        SetLoadFromText(text);
+    }
+}
